requirements.txt
streamlit==1.37.0
pandas==2.2.2
altair==5.3.0
duckdb==1.0.0
python-dotenv==1.0.1
openai==1.40.0
